% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{distance_df}
\alias{distance_df}
\title{Title Distance calculation}
\usage{
distance_df(
  lb_train,
  lb_test_ind,
  match_methods = c("euclidean", "mahalanobis", "single"),
  match_time = NULL,
  id_var,
  outcome_var,
  time_var,
  ...
)
}
\arguments{
\item{lb_train}{the prediction/imputations from the two stage models (brokenstick model and linear model) dataset for the training data}

\item{lb_test_ind}{the prediction/imputations from the two stage models (brokenstick model and linear model) dataset for the testing data, this is the prediction from the test individual}

\item{match_methods}{the distance calculation methods, either "euclidean", "mahalanobis", or "single"}

\item{match_time}{the critical time point for single-time matching, the time must be included in the anchor time set; do not need to specify the time for "euclidean" and "mahalanobis" methods}

\item{id_var}{the id variable in the dataset}

\item{outcome_var}{the outcome variable in the dataset}

\item{time_var}{the time variable in the dataset}

\item{...}{}
}
\value{
a distance dataframe with the distance and p-value arranged
from smallest to largest
}
\description{
Title Distance calculation
}
\examples{
\dontrun{}
}
