% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{euclidean_df}
\alias{euclidean_df}
\title{Title: Calculate Euclidean distance for a matrix}
\usage{
euclidean_df(Dmatrix, center)
}
\arguments{
\item{Dmatrix}{A numeric data matrix}

\item{center}{A numeric vector}
}
\value{
The Euclidean distance list order
from the smallest to the largest difference
}
\description{
A short description...
Calculates the Euclidean distance of a given data matrix,
at an arbitrary center.
}
\examples{
\dontrun {}
}
