% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{euclidean_n}
\alias{euclidean_n}
\title{Title Euclidean distance matching with $\kappa$ criteria}
\usage{
euclidean_n(Dmatrix, match_num)
}
\arguments{
\item{Dmatrix}{the distance matrix for calculating Euclidan distance}

\item{match_num}{the number of matching $\kappa$ criteria}
}
\value{
A subset of distance list arranged from smallest to largest
}
\description{
Title Euclidean distance matching with $\kappa$ criteria
}
\examples{
\dontrun {}
# example code

}
