% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_drizopoulos_code.R
\name{IndvPred_lmer}
\alias{IndvPred_lmer}
\title{Individualized Predictions from Linear Mixed Models for \code{lme4::lmer()}}
\usage{
IndvPred_lmer(
  lmerObject,
  data,
  newdata,
  timeVar,
  outcomeVar,
  idVar,
  times = NULL,
  lmer.fixed = "1 + bs(time, knots = c(5, 10, 15), degree = 1)",
  lmer.random = "1 + bs(time, knots = c(5, 10, 15), degree = 1)",
  M = 200L,
  interval = c("confidence", "prediction"),
  all_times = TRUE,
  level = 0.95,
  return_data = TRUE,
  seed = 1L,
  ...
)
}
\arguments{
\item{lmerObject}{An object of class \code{lmerMod} resulting from \code{lmer()}.}

\item{data}{A data frame used to train the linear mixed model.}

\item{newdata}{A data frame containing the variables for the new subjects
for which predictions are needed.}

\item{timeVar}{A character string specifying the time variable in the data.}

\item{outcomeVar}{A character string specifying the outcome variable.}

\item{idVar}{A character string specifying the ID variable representing individuals.}

\item{times}{A numeric vector specifying the time points for which
subject-specific predictions should be calculated. Defaults to
100 equally spaced time points.}

\item{lmer.fixed}{The formula for the fixed effects part of the model.}

\item{lmer.random}{The formula for the random effects part of the model.}

\item{M}{An integer specifying the number of Monte Carlo samples. Default is 200.}

\item{interval}{A character string indicating the type of intervals to compute:
either "confidence" or "prediction". Defaults to "confidence".}

\item{all_times}{Logical; if \code{TRUE}, predictions are computed at all
time points in \code{times}, otherwise only after the last observed time for each subject.}

\item{level}{A numeric scalar specifying the confidence level for the prediction intervals. Default is 0.95.}

\item{return_data}{Logical; if \code{TRUE}, the function returns the data frame supplied in
\code{newdata} augmented with the predictions and intervals. If \code{FALSE},
a list with predicted values and intervals is returned.}

\item{seed}{An integer used to set the random seed for reproducibility. Default is 1.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A data frame or list with the following components:
\itemize{
\item If \code{return_data = TRUE}, a data frame \code{newdata} with additional rows for
the time points where predictions were made, and additional columns for the predicted values
and confidence/prediction intervals.
\item If \code{return_data = FALSE}, a list with:
\item{times_to_pred}{Time points where predictions were made.}
\item{predicted_y}{The predicted values at these time points.}
\item{low}{The lower bounds of the prediction intervals.}
\item{upp}{The upper bounds of the prediction intervals.}
}
}
\description{
This function provides subject-specific predictions from a fitted
\code{lmerMod} object (created by \code{lme4::lmer()}). It calculates the predictions
using Monte Carlo sampling, optionally generating prediction intervals,
and offers flexibility for predicting at all specified time points or only
after the last observed time for each subject.
}
\note{
This function is adapted from the \code{IndvPred_lme()} function in the \code{JMbayes} package. Please see the \code{JMbayes} package for more details.
}
\examples{
\dontrun{
  # Example with lmer:
  library(lme4)
  lmer_mod <- lmer(y ~ time + (1 | id), data = example_data)

  # Individualized prediction for subject with id = 2:
  pred_result <- IndvPred_lmer(lmerObject = lmer_mod,
                               data = example_data,
                               newdata = subset(example_data, id == 2),
                               timeVar = "time",
                               outcomeVar = "y",
                               idVar = "id",
                               M = 1000,
                               return_data = TRUE)
}
}
