% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{letters_only}
\alias{letters_only}
\title{Title Check whether the object is letters-only}
\usage{
letters_only(x)
}
\arguments{
\item{x}{the string element}
}
\value{
a Boolean result as TRUE or FALSE
}
\description{
Title Check whether the object is letters-only
}
