% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_impute.R
\name{multiple_impute}
\alias{multiple_impute}
\title{Title Multiple Linear Regression imputation at anchor time points}
\usage{
multiple_impute(
  lm_formula,
  data_impute,
  data_test,
  id_var,
  outcome_var,
  time_var,
  anchor_time,
  ...
)
}
\arguments{
\item{lm_formula}{the second stage model with a user defined linear regression formula,
it depends on the liner model type and anchor time points used.}

\item{data_impute}{the design matrix for training data set for the linear regression model,}

\item{data_test}{the design matrix for testing data set for the linear regression model,}

\item{id_var}{the id variable name must be included in the dataset}

\item{outcome_var}{the outcome variable name must be included in the dataset}

\item{time_var}{the time variable name must be included in the dataset}

\item{anchor_time}{the anchor time set for imputation}

\item{...}{Not used, but required for future extension}
}
\value{
A list with the imputed values for the training data set and testing data set,
and the summarization for the linear models probably will be removed in the furture.
}
\description{
The function is used to impute the missing values after the brokenstick model;
the function will take the outcomes from the brokenstick model and impute the outcomes at the
anchor time points using multiple linear regressions model at each anchor time point.
We will include extra time-invariant covariates. Different from \code{linear_impute()} function,
Here we fit multiple linear regression models at each anchor time point independently,
which grant more flexibility for the models.
This function is mainly used in \code{people-like-me()} and \code{people-like-us()} function,
as an alternate to \code{linear_impute()}.
}
\examples{
\dontrun {}
}
