% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{not_any_na}
\alias{not_any_na}
\title{Title Checking there is any not NA}
\usage{
not_any_na(x)
}
\arguments{
\item{x}{the target dataset}
}
\value{
a boolean result
}
\description{
Title Checking there is any not NA
}
