% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{numbers_only}
\alias{numbers_only}
\title{Title Check whether the object is number-only}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{the string element}
}
\value{
a Boolean result as TRUE or FALSE
}
\description{
Title Check whether the object is number-only
}
