% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_people-like-us.R
\name{people_like_i}
\alias{people_like_i}
\title{People-Like-Me for Single-Time Point Matching for Single Individual}
\usage{
people_like_i(
  train_data,
  test_data,
  outcome_var,
  time_var,
  id_var,
  tmin,
  tmax,
  brokenstick_knots,
  anchor_time,
  linear_formula,
  gamlss_formula,
  gamlss_sigma,
  match_number,
  match_plot = TRUE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{train_data}{\code{data.frame} The training dataset for fitting the brokenstick and linear models. This dataset also serves as the pool for the matching process.}

\item{test_data}{\code{data.frame} The testing dataset for personalized prediction.}

\item{outcome_var}{\code{character} The name of the outcome variable of interest.}

\item{time_var}{\code{character} The name of the time variable of interest.}

\item{id_var}{\code{character} The name of the ID variable representing individuals.}

\item{tmin}{\code{numeric} The minimum time point for the prediction.}

\item{tmax}{\code{numeric} The maximum time point for the prediction.}

\item{brokenstick_knots}{\code{numeric} A vector of knots for the brokenstick model, which do not need to be equally spaced.}

\item{anchor_time}{\code{numeric} The specific time point used as the anchor for the matching process.}

\item{linear_formula}{\code{formula} The formula used to fit the linear model.}

\item{gamlss_formula}{\code{formula} The mean formula for the GAMLSS model, typically used for smoothing as a function of time.}

\item{gamlss_sigma}{\code{formula} The sigma formula for the GAMLSS model, used for smoothing as a function of time.}

\item{match_number}{\code{numeric} The number of matches to find during the matching process. Can be \code{NULL} or a numeric value.}

\item{match_plot}{\code{logical} If \code{TRUE}, a plot of the matching process will be generated. Defaults to \code{TRUE}.}

\item{predict_plot}{\code{logical} If \code{TRUE}, a plot of the prediction process will be generated. Defaults to \code{TRUE}.}

\item{...}{Other arguments passed to the function.}
}
\value{
A \code{list} containing the following attributes:
\itemize{
\item \code{predicted_centiles}: A \code{data.frame} of predicted centiles for each individual in the test dataset.
\item \code{observed_centiles}: A \code{data.frame} of observed centiles for each individual in the test dataset.
\item \code{plots}: A \code{ggplot} object of the prediction plot, if \code{predict_plot = TRUE}.
\item \code{subset}: A list of matched individuals based on the single-time point matching process.
\item \code{linear}: The fitted linear model used to predict outcomes.
\item \code{brokenstick_model}: The fitted brokenstick model used for initial data imputation and prediction.
\item \code{matching_plot}: A plot object of the matching process, if \code{match_plot = TRUE}.
}
}
\description{
This function, part of the People-Like-Me method, predicts personalized centiles for a testing dataset based on a training dataset. It first fits a brokenstick model and a linear model to the training data, then predicts outcomes for the test data using the linear model. Matches for the test dataset are determined based on single time point predictions. Finally, the function predicts centiles for the test dataset using a GAMLSS model, with optional plots if \code{predict_plot} is set to \code{TRUE}.
}
\examples{
\dontrun{
  # Example usage
  results <- people_like_i(train_data, test_data, outcome_var = "height", time_var = "age", id_var = "id", ...)
}

}
\seealso{
\code{\link{people_like_me}}, \code{\link{people_like_us}}, \code{\link{people_like_thee}}
}
