% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_people-like-me.R
\name{people_like_me}
\alias{people_like_me}
\title{Title People-Like-Me methods for single testing individual}
\usage{
people_like_me(
  train_data,
  test_data,
  outcome_var,
  time_var,
  id_var,
  tmin,
  tmax,
  brokenstick_knots,
  anchor_time,
  linear_formula,
  gamlss_formula,
  gamlss_sigma,
  match_methods = c("euclidean", "mahalanobis", "single"),
  weight = FALSE,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = TRUE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title People-Like-Me methods for single testing individual
}
