% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_people-like-thee.R
\name{people_like_thee}
\alias{people_like_thee}
\title{Title People-Like-Me for Dynamic Prediction}
\usage{
people_like_thee(
  train_data,
  test_data,
  new_data,
  outcome_var = "ht",
  time_var = "time",
  id_var = "id",
  tmin,
  tmax,
  brokenstick_knots,
  anchor_time = "(5, 10, 15)",
  gamlss_formula,
  gamlss_sigma,
  match_methods = c("euclidean", "mahalanobis"),
  weight = FALSE,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = FALSE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{train_data}{the training dataset}

\item{test_data}{the testing dataset}

\item{new_data}{the new dataset for prediction}

\item{outcome_var}{the outcome variable in the dataset}

\item{time_var}{the time variable in the dataset}

\item{id_var}{the id variable in the dataset}

\item{tmin}{the minimum time for the prediction}

\item{tmax}{the maximum time for the prediction}

\item{brokenstick_knots}{the knots for the brokenstick model}

\item{anchor_time}{the anchor time for the prediction}

\item{gamlss_formula}{the formula for the GAMLSS model mean function part}

\item{gamlss_sigma}{the formula for the GAMLSS model sigma function part}

\item{match_methods}{the methods for matching the individuals}

\item{weight}{the weighted regression inputs for the GAMLSS model}

\item{match_alpha}{the alpha level for the matching}

\item{match_number}{the number of matches for the matching}

\item{match_plot}{whether you want to include the plot for the matching}

\item{predict_plot}{whether you want to include the plot for the prediction}

\item{...}{other arguments}
}
\value{
a list of predicted values for the testing dataset with observational time, the predictive time, and the predicted values.
}
\description{
Title People-Like-Me for Dynamic Prediction
}
\examples{
\dontrun{}
}
