% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{predict_gamlss}
\alias{predict_gamlss}
\title{Prediction with GAMLSS Model}
\usage{
predict_gamlss(
  matching,
  test_one,
  id_var,
  time_var,
  outcome_var,
  tmin,
  tmax,
  gamlss_formula,
  gamsigma_formula,
  weight = FALSE,
  predict_plot
)
}
\arguments{
\item{matching}{\code{data.frame} The dataset of matched individuals from the original dataset.}

\item{test_one}{\code{data.frame} The dataset for the target individual for whom predictions are made.}

\item{id_var}{\code{character} The name of the ID variable in the dataset.}

\item{time_var}{\code{character} The name of the time variable in the dataset.}

\item{outcome_var}{\code{character} The name of the outcome variable in the dataset.}

\item{tmin}{\code{numeric} The minimum time point for the prediction.}

\item{tmax}{\code{numeric} The maximum time point for the prediction.}

\item{gamlss_formula}{\code{formula} The formula for the mean function of the GAMLSS model.}

\item{gamsigma_formula}{\code{formula} The formula for the sigma (scale) function of the GAMLSS model.}

\item{weight}{\code{logical} Indicates whether to use weighted regression for the GAMLSS model based on matching. Defaults to \code{FALSE}.}

\item{predict_plot}{\code{logical} Whether to include a plot of the predictions. Defaults to \code{TRUE}.}
}
\value{
A list containing:
\itemize{
\item \code{centiles_observed}: A \code{data.frame} of observed centiles for the individual, along with coverage and bias.
\item \code{centiles_predicted}: A \code{data.frame} of predicted centiles (5th, 10th, 25th, 50th, 75th, 90th, 95th percentiles) over time.
\item \code{predictive_centiles}: A \code{ggplot} object of the centile prediction plot if \code{predict_plot = TRUE}, otherwise \code{NULL}.
}
}
\description{
This function performs centile prediction using the GAMLSS model for a target individual. It fits the GAMLSS model to the matching dataset and predicts the centiles for the test individual. The function can optionally generate a plot comparing the predicted centiles with the individual’s observational data.
}
\examples{
\dontrun{
  result <- predict_gamlss(matching = matched_data,
                           test_one = individual_data,
                           id_var = "id",
                           time_var = "time",
                           outcome_var = "height",
                           tmin = 1, tmax = 10,
                           gamlss_formula = "height ~ time",
                           gamsigma_formula = "~ time",
                           weight = FALSE,
                           predict_plot = TRUE)
}

}
\seealso{
\code{\link{plm_ind_plot}} for the plotting function used.
}
