% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{singletime_n}
\alias{singletime_n}
\title{Title Single time point matching with both matching time and matching number}
\usage{
singletime_n(Dmatrix, match_time, match_num)
}
\arguments{
\item{Dmatrix}{A distance matrix}

\item{match_time}{The matching anchor time as setup}

\item{match_num}{The number of matching $\kappa$ criteria}
}
\value{
A distance list arranged from smallest to largest
}
\description{
Calculate the distance of a single time point
with a given anchor time point and the matching number
}
\examples{
\dontrun {}
}
