% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dataset.R
\docType{data}
\name{train}
\alias{train}
\title{Example dataset for EPIC children height growth simulated based on Preece-Baine I Model}
\format{
A training tibble with 9363 rows and 5 variables and
a testing tibble with 186 rows and 5 variables:
\describe{
\item{id}{fct}{the individual identifier was recorded}
\item{time}{dbl}{the time of measurement}
\item{sex}{fct}{the gender of the given individual}
\item{ht}{dbl}{Heights for the Individual}
\item{group}{fct}{training or Testing group}
}
}
\source{
\url{}
}
\usage{
train
}
\description{
This an example set for the EPIC children height growth dataset. The dataset
contains 500 individuals for training set and 10 individuals for testing set
The variables are:
}
\keyword{datasets}
