% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{dis_match}
\alias{dis_match}
\title{Title The distance calculation function for the PLM}
\usage{
dis_match(
  lb_train,
  lb_test_ind,
  train = train,
  match_methods = c("euclidean", "mahalanobis", "single"),
  id_var,
  outcome_var,
  time_var,
  match_alpha = NULL,
  match_number = NULL,
  match_time = NULL,
  match_plot,
  ...
)
}
\arguments{
\item{lb_train}{the prediction/imputations from the two stage models (brokenstick model and linear model) dataset for the training data}

\item{lb_test_ind}{the prediction/imputations from the two stage models (brokenstick model and linear model) dataset for the testing data, this is the prediction from the test individual}

\item{train}{the training dataset because we need to subset the training dataset to get the matched individuals}

\item{match_methods}{the distance calculation methods, either "euclidean", "mahalanobis", or "single"}

\item{id_var}{the id variable in the dataset}

\item{outcome_var}{the outcome variable in the dataset}

\item{time_var}{the time variable in the dataset}

\item{match_alpha}{the p-value threshold for the matching (specifically for the Mahalanobis distance matching PLM)}

\item{match_number}{the number of matches you want to get from the distance dataframe (specifically for the Euclidean and Mahalanobis distance matching PLM)}

\item{match_time}{the critical time point for single-time matching, the time must be included in the anchor time set; do not need to specify the time for "euclidean" and "mahalanobis" methods}

\item{match_plot}{whether you want to plot the matching individual trajectories}

\item{...}{other arguments for the function}
}
\value{
a list with the matching subset, the plot of the matching individual trajectories, the id of the target individual, the alpha threshold for the Mahalanobis distance matching, and the number of matches for the Euclidean and Mahalanobis distance matching
}
\description{
Title The distance calculation function for the PLM
}
\examples{
\dontrun{}

}
