% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{distance_df}
\alias{distance_df}
\title{Title Distance calculation}
\usage{
distance_df(
  lb_train,
  lb_test_ind,
  match_methods = c("euclidean", "mahalanobis", "single"),
  match_time = NULL,
  id_var,
  outcome_var,
  time_var,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title Distance calculation
}
