% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{euclidean_df}
\alias{euclidean_df}
\title{Calculate Euclidean Distance for a Matrix}
\usage{
euclidean_df(Dmatrix, center)
}
\arguments{
\item{Dmatrix}{\code{matrix} A numeric data matrix where each column represents a data point.}

\item{center}{\code{numeric} A numeric vector representing the center point from which the Euclidean distances are calculated.}
}
\value{
A \code{data.frame} with two columns: \code{id}, which represents the column names of the data matrix, and \code{diff}, which contains the Euclidean distances, ordered from smallest to largest.
}
\description{
This function calculates the Euclidean distance between each column in a given numeric data matrix and a specified center. The result is returned as a list of distances, ordered from smallest to largest difference.
}
\examples{
\dontrun{
  # Example usage
  Dmatrix <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
  center <- c(1, 2)
  euclidean_df(Dmatrix, center)
}
}
