% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{euclidean_n}
\alias{euclidean_n}
\title{Euclidean Distance Matching with \eqn{\kappa} Criteria}
\usage{
euclidean_n(Dmatrix, match_num)
}
\arguments{
\item{Dmatrix}{\code{matrix} A numeric distance matrix where columns represent data points.}

\item{match_num}{\code{numeric} The number of closest matches to return, based on Euclidean distance.}
}
\value{
A \code{data.frame} with two columns: \code{id}, representing the identifiers for each column in the matrix, and \code{diff}, which contains the calculated Euclidean distances, ordered from smallest to largest. The result is limited to the top \code{match_num} closest matches.
}
\description{
This function calculates the Euclidean distance for each column in a given distance matrix and returns the top \code{match_num} closest matches based on the smallest Euclidean distances.
}
\examples{
\dontrun{
  # Example usage
  Dmatrix <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3)
  match_num <- 2
  euclidean_n(Dmatrix, match_num)
}
}
