% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_impute.R
\name{impute_brokenstick}
\alias{impute_brokenstick}
\title{Title First Stage model imputation with \code{brokenstick::brokenstick()} at anchor time points}
\usage{
impute_brokenstick(
  outcome_var,
  time_var,
  id_var,
  bs_knots,
  anchor_time,
  data,
  ...
)
}
\arguments{
\item{outcome_var}{The outcome_var variable name must include in the dataset}

\item{time_var}{The time variable name must be included in the dataset}

\item{id_var}{The id variable name must be included in the dataset}

\item{bs_knots}{The internal knots for brokenstick model}

\item{anchor_time}{The anchor time set for imputation,}

\item{data}{A data frame in which to look for variables with which to fit
the brokenstick model and predict. Ideally, this is a longitudinal
data.frame object in long-format.}

\item{...}{Not used, but required for future extension}
}
\value{
A data frame with the imputed values
}
\description{
The predictions from a broken stick model coincide with
the group-conditional means of the random effects.
This function takes the \code{data} with selected \code{outcome_var}, \code{time_var},
and \code{id_var}. The user can calculate prediction (imputation) at
given anchor \code{time} set.
}
\examples{
\dontrun {}
}
