% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_drizopoulos_code.R
\name{IndvPred_lmer}
\alias{IndvPred_lmer}
\title{Title Individualized Predictions from Linear Mixed Models for \code{lme4::lmer()}
(Adapted from Dimitris Rizopoulos)}
\usage{
IndvPred_lmer(
  lmerObject,
  data,
  newdata,
  timeVar,
  outcomeVar,
  idVar,
  times = NULL,
  lmer.fixed = "1 + bs(time, knots = c(5, 10, 15), degree = 1)",
  lmer.random = "1 + bs(time, knots = c(5, 10, 15), degree = 1)",
  M = 200L,
  interval = c("confidence", "prediction"),
  all_times = TRUE,
  level = 0.95,
  return_data = TRUE,
  seed = 1L,
  ...
)
}
\arguments{
\item{lmerObject}{an object inheriting from function \code{lmer()},
class must be \code{lmerMod}.}

\item{data}{a data frame in which to train the linear mixed model
(need to be updated in the plm methods).}

\item{newdata}{a data frame in which to look for variables with which to predict.}

\item{timeVar}{a character string indicating what type of intervals should be computed.}

\item{times}{a numeric vector denoting the time points for which we wish to compute the
subject-specific predictions after the last available measurement provided in
\code{newdata}. Default is a sequence of 100 equally spaced time points
from the smallest to the largest follow-up time of all subjects.}

\item{M}{numeric scalar denoting the number of Monte Carlo samples.}

\item{interval}{a character string indicating what type of intervals should be computed.}

\item{all_times}{logical; should predictions be calculated at all \code{times} or only
at the ones that are after the last observed time of each subject.}

\item{level}{a numeric scalar denoting the tolerance/confidence level.}

\item{return_data}{logical; if \code{TRUE} the data frame supplied in
\code{newdata} is returned augmented with the outputs of the function.}

\item{seed}{numeric scalar, the random seed used to produce the results.}
}
\value{
a data frame with the predicted values,
lower and upper bounds of the prediction intervals.

If \code{return_data = TRUE},
a the data frame \code{newdata} with extra rows for the time points at
which predictions were calculated, and extra columns with the predictions
and the limits of the pointwise confidence intervals.

If \code{return_data = FALSE}, a list with components
\item{times_to_pred}{time points at which predictions were calculated.}
\item{predicted_y}{the predictions.}
\item{low}{the lower limits of the pointwise confidence intervals.}
\item{upp}{the upper limits of the pointwise confidence intervals.}
}
\description{
Title Individualized Predictions from Linear Mixed Models for \code{lme4::lmer()}
(Adapted from Dimitris Rizopoulos)
}
\examples{
{
\dontrun{
# linear mixed model fit
 lmer <- lmer()

 id2 <- IndvPred_lmer(lmer,
                      newdata = subset(tsa_test1, id == 2),
                      timeVar = "time",
                      M = 1000,
                      return_data = TRUE)
  }
}
}
