% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_impute.R
\name{linear_impute}
\alias{linear_impute}
\title{Title Linear regression for people-like-me methods}
\usage{
linear_impute(
  lm_formula,
  data_impute,
  data_test,
  id_var,
  outcome_var,
  time_var,
  anchor_time,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title Linear regression for people-like-me methods
}
