% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{mahalanobis_n}
\alias{mahalanobis_n}
\title{Title Mahalanobis distance with $\kappa$ criteria}
\usage{
mahalanobis_n(Dmatrix, match_num)
}
\arguments{
\item{Dmatrix}{the distance matrix for calculating Mahalanobis distance}

\item{match_num}{the number of matching $\kappa$ criteria}
}
\value{
A subset of Mahalanobis distance list arranged from smallest to largest
}
\description{
Title Mahalanobis distance with $\kappa$ criteria
}
\examples{
\dontrun {}
}
