% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{mahalanobis_p}
\alias{mahalanobis_p}
\title{Title Mahalanobis distance with p-value $\alpha$ criteria}
\usage{
mahalanobis_p(Dmatrix, alpha)
}
\arguments{
\item{Dmatrix}{the distance matrix for calculating Mahalanobis distance}

\item{alpha}{the p-value criteria}
}
\value{
A subset of Mahalanobis distance list arranged from smallest to largest
}
\description{
Title Mahalanobis distance with p-value $\alpha$ criteria
}
\examples{
\dontrun {}
}
