% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{match}
\alias{match}
\title{Title Finding the matches subset}
\usage{
match(
  distance_df = ddd,
  train = train,
  test_one,
  id_var,
  outcome_var,
  time_var,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = FALSE
)
}
\arguments{
\item{match_plot}{}
}
\description{
Title Finding the matches subset
}
