% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{match}
\alias{match}
\title{Finding the Matches Subset}
\usage{
match(
  distance_df = ddd,
  train = train,
  test_one,
  id_var,
  outcome_var,
  time_var,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = FALSE
)
}
\arguments{
\item{distance_df}{\code{data.frame} The distance dataframe containing calculated distances (and p-values, if applicable), typically arranged from smallest to largest.}

\item{train}{\code{data.frame} The training dataset to subset based on the matching process.}

\item{test_one}{\code{data.frame} The testing individual’s data for whom matching is performed.}

\item{id_var}{\code{character} The name of the ID variable in the dataset.}

\item{outcome_var}{\code{character} The name of the outcome variable in the dataset.}

\item{time_var}{\code{character} The name of the time variable in the dataset.}

\item{match_alpha}{\code{numeric} The p-value threshold for Mahalanobis distance matching. If \code{NULL}, it is ignored.}

\item{match_number}{\code{numeric} The number of matches to retrieve from the distance dataframe, specifically for Euclidean and Mahalanobis distance matching. If \code{NULL}, it is ignored.}

\item{match_plot}{\code{logical} Whether to generate a plot of the matching individuals’ trajectories. Defaults to \code{FALSE}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{subset}: A \code{data.frame} containing the matched subset from the training dataset.
\item \code{plot}: A \code{ggplot} object showing the trajectories of the matching individuals compared to the test individual if \code{match_plot = TRUE}, otherwise \code{NULL}.
\item \code{id}: The ID of the target test individual.
\item \code{alpha}: The alpha threshold used for Mahalanobis distance matching, or \code{NULL} if not used.
\item \code{number}: The number of matches used for Euclidean or Mahalanobis distance matching, or \code{NULL} if not used.
}
}
\description{
This function finds the matching individuals from the training dataset based on a distance dataframe, such as one generated by the \code{distance_df()} function. The function allows for filtering by p-value or selecting a specified number of matches, and can optionally generate a plot comparing the matching individual trajectories with the test individual.
}
\examples{
\dontrun{
  matches <- match(distance_df = distance_data,
                   train = train_data,
                   test_one = test_individual_data,
                   id_var = "id",
                   outcome_var = "height",
                   time_var = "age",
                   match_alpha = 0.95,
                   match_plot = TRUE)
}
}
