% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{not_all_na}
\alias{not_all_na}
\title{Title Checking there is not all NA}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{the target dataset}
}
\value{
a boolean result
}
\description{
Title Checking there is not all NA
}
