% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_people-like-us.R
\name{people_like_i}
\alias{people_like_i}
\title{Title Original People-Like-Me for Single-time Matching}
\usage{
people_like_i(
  train_data,
  test_data,
  outcome_var,
  time_var,
  id_var,
  tmin,
  tmax,
  brokenstick_knots,
  anchor_time,
  linear_formula,
  gamlss_formula,
  gamlss_sigma,
  match_number,
  match_plot = TRUE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{train_data}{The training dataset for brokenstick model and linear model fitting; This training dataset also serves as the pool for matching process.}

\item{test_data}{The testing dataset for the personalized prediction.}

\item{outcome_var}{The outcome variable of interest.}

\item{time_var}{The time variable of interest.}

\item{id_var}{The id variable of interest.}

\item{tmin}{The minimum time point for the prediction.}

\item{tmax}{The maximum time point for the prediction.}

\item{brokenstick_knots}{The knots for the brokenstick model, which does not need to be equal distanced.}

\item{anchor_time}{The time point for the anchor time for the matching process.}

\item{linear_formula}{The formula for the linear model.}

\item{gamlss_formula}{The mean formula for the GAMLSS model, mainly used as a smoothing process only include the function of time. Here we use the GAMLSS model as a non-parametric / semi-parametric functional process.}

\item{gamlss_sigma}{The sigma formula for the GAMLSS model, mainly used as a smoothing process only include the function of time. Here we use the GAMLSS model as a non-parametric / semi-parametric functional process.}

\item{match_number}{The number of matches for the matching process, which can be NULL for no number of matches, or a numeric value for the number of matches.}

\item{match_plot}{The logical value for the matching plot, which can be TRUE for the plot, FALSE for no plot.}

\item{predict_plot}{The logical value for the prediction plot, which can be TRUE for the plot, FALSE for no plot.}

\item{...}{Other arguments for the function.}
}
\value{
A list of the predicted centiles, observed centiles, and the plot if the \code{predict_plot} argument is TRUE.
}
\description{
The People-Like-Me fucntion is to predict the personalized centiles for the testing dataset based on the training dataset. The function will first fit the brokenstick model and linear model for the training dataset. Then, the function will predict the outcome for the testing dataset based on the linear model. The function will find the matches for the testing dataset based on the single time point prediction. The function will predict the centiles for the testing dataset based on the GAMLSS model and the plot if the \code{predict_plot} argument is TRUE. This function is the basic setup for the People-Like-Me method with single time point matching.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link[=people_like_me]{people_like_me()}}, \code{\link[=people_like_us]{people_like_us()}}, \code{\link[=people_like_thee]{people_like_thee()}}
}
