% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_people-like-us.R
\name{people_like_me}
\alias{people_like_me}
\title{Title People-Like-Me methods for single testing individual}
\usage{
people_like_me(
  train_data,
  test_data,
  outcome_var,
  time_var,
  id_var,
  tmin,
  tmax,
  brokenstick_knots,
  anchor_time,
  linear_formula,
  gamlss_formula,
  gamlss_sigma,
  match_methods = c("euclidean", "mahalanobis", "single"),
  weight = FALSE,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = TRUE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{train_data}{The training dataset for brokenstick model and linear model fitting; This training dataset also serves as the pool for matching process.}

\item{test_data}{The testing dataset for the personalized prediction.}

\item{outcome_var}{The outcome variable of interest.}

\item{time_var}{The time variable of interest.}

\item{id_var}{The id variable of interest.}

\item{tmin}{The minimum time point for the prediction.}

\item{tmax}{The maximum time point for the prediction.}

\item{brokenstick_knots}{The knots for the brokenstick model, which does not need to be equal distanced.}

\item{anchor_time}{The time point for the anchor time for the matching process.}

\item{linear_formula}{The formula for the linear model.}

\item{gamlss_formula}{The mean formula for the GAMLSS model, mainly used as a smoothing process only include the function of time. Here we use the GAMLSS model as a non-parametric / semi-parametric functional process.}

\item{gamlss_sigma}{The sigma formula for the GAMLSS model, mainly used as a smoothing process only include the function of time. Here we use the GAMLSS model as a non-parametric / semi-parametric functional process.}

\item{match_number}{The number of matches for the matching process, which can be NULL for no number of matches, or a numeric value for the number of matches.}

\item{match_plot}{The logical value for the matching plot, which can be TRUE for the plot, FALSE for no plot.}

\item{predict_plot}{The logical value for the prediction plot, which can be TRUE for the plot, FALSE for no plot.}

\item{...}{Other arguments for the function.}
}
\value{
A list of the predicted centiles, observed centiles, and the plot if the \code{predict_plot} argument is TRUE.
}
\description{
Title People-Like-Me methods for single testing individual
}
\examples{
\dontrun{}
}
