% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{plm_ind_plot}
\alias{plm_ind_plot}
\title{Title plot individual matching}
\usage{
plm_ind_plot(
  quantile,
  observation,
  title = NULL,
  outcome_var,
  id_var,
  time_var,
  ...
)
}
\arguments{
\item{quantile}{the quantiles you want to plot out with the prediction}

\item{observation}{the observation data from the original dataset}

\item{title}{the title for the plots}

\item{outcome_var}{the outcome variable in the dataset}

\item{id_var}{the id variable in the dataset}

\item{time_var}{the time variable in the dataset}
}
\value{
a ggplot object the plot with target individual observational data
and all the matches for the target predictions
}
\description{
Title plot individual matching
}
