% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{plm_ind_plot}
\alias{plm_ind_plot}
\title{Title plot individual matching}
\usage{
plm_ind_plot(
  quantile,
  observation,
  title = NULL,
  outcome_var,
  id_var,
  time_var,
  ...
)
}
\arguments{
\item{title}{}
}
\description{
Title plot individual matching
}
