% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{predict_gamlss}
\alias{predict_gamlss}
\title{Title Prediction with GAMLSS model}
\usage{
predict_gamlss(
  matching,
  test_one,
  id_var,
  time_var,
  outcome_var,
  tmin,
  tmax,
  gamlss_formula,
  gamsigma_formula,
  weight = FALSE,
  predict_plot
)
}
\arguments{
\item{predict_plot}{}
}
\description{
Title Prediction with GAMLSS model
}
