% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Title Read Multiple Sheets of Excel file}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{the directory location of the excel file}

\item{tibble}{return the datasets in as.data.frame format}
}
\value{
a list of data.frames
}
\description{
Title Read Multiple Sheets of Excel file
}
