% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{single_df}
\alias{single_df}
\title{Single Time Point Matching}
\usage{
single_df(Dmatrix, match_time, center)
}
\arguments{
\item{Dmatrix}{\code{matrix} A numeric distance matrix where rows represent time points.}

\item{match_time}{\code{numeric} The anchor time point for matching. This specifies which row (time point) to use for distance calculation.}

\item{center}{\code{numeric} The value representing the center (reference point) from which distances are calculated.}
}
\value{
A \code{data.frame} with two columns: \code{id}, which represents the identifiers for each element, and \code{diff}, which contains the calculated distances, ordered from smallest to largest absolute difference from the center.
}
\description{
This function calculates the distance at a single time point from a given distance matrix. It returns a list of distances, ordered from smallest to largest based on the absolute difference from a specified center.
}
\examples{
\dontrun{
  # Example usage
  Dmatrix <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3)
  match_time <- 2
  center <- 3
  single_df(Dmatrix, match_time, center)
}
}
